+++ src/helpers/presence.rs
@@ -1,5 +1,5 @@
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
-/// Defines wether a claim is required or not
+/// Defines whether a claim is required or not
 pub enum Presence {
     /// Claim is optional
     Optional,
